+++ src/main/java/com/ford/afd/repository/FoundationDataColumnRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.ford.afd.model.FoundationDataColumn;
 
-public interface FoundationDataColumnRepository extends JpaRepository<FoundationDataColumn, Long> {
+public interface FoundationDataColumnRepository extends JpaRepository<FoundationDataColumn, Integer> {
 	
 	/**
 	 * Checks if uiColumnName already existing in FoundationDataColumn table..
