+++ src/main/java/com/ford/afd/repository/MasterDataItemRepository.java
@@ -6,14 +6,14 @@ import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MasterDataItemRepository extends JpaRepository<MasterDataItem, Long> {
+public interface MasterDataItemRepository extends JpaRepository<MasterDataItem, Integer> {
 	
 	/**
 	 * Retrieves MasterDataItem for given masterDataId.
 	 * @param masterDataId
 	 * @return
 	 */
-	List<MasterDataItem> findByParentMasterDataId(long masterDataId);
+	List<MasterDataItem> findByParentMasterDataId(int masterDataId);
 	
 	/**
 	 * Checks if code already existing in MasterDataItem table.
