+++ src/main/java/com/ford/afd/repository/MasterDataRepository.java
@@ -3,7 +3,7 @@ package com.ford.afd.repository;
 import com.ford.afd.model.MasterData;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MasterDataRepository extends JpaRepository<MasterData, Long> {
+public interface MasterDataRepository extends JpaRepository<MasterData, Integer> {
 	
 	/**
 	 * Checks if name already existing in MasterData table..
